2023.06b

+++ Pico/Trump/Melania/fish0002a.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
+#include <cstdint>
 
-void foo(const int *x, int count){
+void foo(const uint8_t *x, int count){
 	for (int i=0; i<count; i++) {
 		std::cout << x[i] << "\n";
 	}
@@ -8,8 +9,8 @@ void foo(const int *x, int count){
 
 int main(void)
 {
-	int x[3] = {3, 2, 6};
+	uint8_t x[3] = {3, 2, 6};
 	foo(x, 3);
 	std::cout << "****\n";
-	foo(  (const int [3]){4, 5, 7}, 3             );
+	foo(  (const uint8_t [3]){4, 5, 7}, 3             );
 }
